fix(backend): resolve PHPStan type errors
Add type annotation for $path array in RouteCalculator and cast
strtotime result to int in PeriodCalculator to satisfy strict typing.

+++ backend/src/Infrastructure/Persistence/PeriodCalculator.php
@@ -22,7 +22,7 @@ class PeriodCalculator
             'day' => [$group, $group],
             'month' => [
                 $group . '-01',
-                date('Y-m-t', strtotime($group . '-01')),
+                date('Y-m-t', (int) strtotime($group . '-01')),
             ],
             'year' => [
                 $group . '-01-01',
